Implement legacy code cleanup for 64-bit Pi 5 migration
## Critical Fixes
- **rotate.cpp**: Replace unaligned pointer cast with memcpy() to prevent ARM64 crashes
- **video_loopback.cpp**: Replace unsafe atoi() with strtol() + error handling (errno checking)

## FFmpeg Modernization
- **netcam.cpp**: Remove deprecated refcounted_frames option (3 lines, no-op in FFmpeg 5+)
- **util.hpp**: Simplify FFmpeg version shims; use modern FFmpeg 6.0+ APIs directly

## Code Quality Improvements
- **camera.cpp**: Remove duplicate movie_norm initialization
- **movie.cpp**: Fix error message (av_copy_packet → av_packet_ref)
- **video_convert.cpp**: Change long int → int for image dimensions
- **jpegutils.cpp**: Change non-standard ulong → unsigned long

## Testing & Deployment
✅ Builds cleanly on Pi 5 (FFmpeg 6.1, libcamera 0.6.0)
✅ Builds cleanly on Pi 4 (FFmpeg 6.1, libcamera 0.6.0)
✅ Functional testing with Camera Module v3 (Pi 5) and v2 (Pi 4)
✅ Deployed to both systems via make install
✅ MotionEye services verified running with updated binaries

8 files modified, 0 files added/deleted
Net change: 32 insertions, 30 deletions

+++ src/video_convert.cpp
@@ -186,9 +186,9 @@ int cls_convert::sonix_decompress(u_char *img_dst, u_char *img_src)
  */
 void cls_convert::bayer2rgb24(u_char *img_dst, u_char *img_src)
 {
-    long int i;
+    int i;
     u_char *rawpt, *scanpt;
-    long int size;
+    int size;  /* width * height fits in int for any practical resolution */
 
     rawpt = img_src;
     scanpt = img_dst;
